5.1.3 Перенесет функции для рендеринга из utils.js в utils/render.js, а остальные функции - в utils/common.js
